+++ lib/src/rules/unnecessary_getters_setters.dart
@@ -89,21 +89,20 @@ class _Visitor extends SimpleAstVisitor<void> {
     }
 
     // Only select getters with setter pairs
-    getters.keys
-        .where((id) => setters.keys.contains(id))
-        .forEach((id) => _visitGetterSetter(getters[id], setters[id]));
+    for (var id in getters.keys) {
+      _visitGetterSetter(getters[id]!, setters[id]);
+    }
   }
 
-  void _visitGetterSetter(
-      MethodDeclaration? getter, MethodDeclaration? setter) {
-    if (getter != null &&
-        setter != null &&
-        isSimpleSetter(setter) &&
+  void _visitGetterSetter(MethodDeclaration getter, MethodDeclaration? setter) {
+    if (setter == null) return;
+    var getterElement = getter.declaredElement;
+    var setterElement = setter.declaredElement;
+    if (getterElement == null || setterElement == null) return;
+    if (isSimpleSetter(setter) &&
         isSimpleGetter(getter) &&
-        !isProtected(getter) &&
-        !isProtected(setter) &&
-        !isDeprecated(getter) &&
-        !isDeprecated(setter)) {
+        getterElement.metadata.isEmpty &&
+        setterElement.metadata.isEmpty) {
       rule..reportLint(getter.name)..reportLint(setter.name);
     }
   }